+++ SLUGS/__init__.py
@@ -1,2 +1,2 @@
-__version__ = '2.1.1'
+__version__ = '2.1.2IP'
 VERSION = __version__  # synonym