+++ SLUGS/urls.py
@@ -36,7 +36,7 @@ urlpatterns = [
         name="download_form",
     ),
     path(
-        "media/estimates/<path:relative_path>",
+        "media/<path:relative_path>",
         EstimateDownload.as_view(),
         name="download_estimate",
     ),